Clean BaseRecyclerFragment & Remove dependancy of 'org.dashj:dashj-core:
1) Move logic  "imgEmptyView.setOnClickListener " from BaseRecyclerFragment
2) Remove dependancy of 'org.dashj:dashj-core:0.14.4.3' from gradle.
