+++ app/src/main/java/pivx/org/pivxwallet/ui/wallet_activity/TransactionsFragmentBase.java
@@ -25,6 +25,7 @@ import pivx.org.pivxwallet.ui.base.tools.adapter.BaseRecyclerAdapter;
 import pivx.org.pivxwallet.ui.base.tools.adapter.BaseRecyclerViewHolder;
 import pivx.org.pivxwallet.ui.base.tools.adapter.ListItemListeners;
 import pivx.org.pivxwallet.ui.transaction_detail_activity.TransactionDetailActivity;
+import pivx.org.pivxwallet.wallofcoins.buyingwizard.BuyDashBaseActivity;
 
 import static pivx.org.pivxwallet.ui.transaction_detail_activity.FragmentTxDetail.IS_DETAIL;
 import static pivx.org.pivxwallet.ui.transaction_detail_activity.FragmentTxDetail.TX_WRAPPER;
@@ -50,6 +51,20 @@ public class TransactionsFragmentBase extends BaseRecyclerFragment<TransactionWr
 
         //setEmptyText("You don't have any transfers yet.");
         setEmptyTextColor(Color.parseColor("#cccccc"));
+        imgEmptyView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getActivity(), BuyDashBaseActivity.class);
+                startActivity(intent);
+            }
+        });
+        txt_empty.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getActivity(), BuyDashBaseActivity.class);
+                startActivity(intent);
+            }
+        });
         return view;
     }
