+++ app/src/main/java/pivx/org/pivxwallet/wallofcoins/utils/MonetarySpannable.java
@@ -23,8 +23,9 @@ import android.text.SpannableString;
 import android.text.style.RelativeSizeSpan;
 import android.text.style.StyleSpan;
 
-import org.bitcoinj.core.Monetary;
-import org.bitcoinj.utils.MonetaryFormat;
+
+import org.pivxj.core.Monetary;
+import org.pivxj.utils.MonetaryFormat;
 
 import java.util.regex.Matcher;
